+++ src/TrendyolClient.Sharp/Models/TrendyolProduct.cs
@@ -27,10 +27,10 @@ public class TrendyolProduct
     public string StockCode { get; set; }
 
     [JsonPropertyName("listPrice")]
-    public double ListPrice { get; set; }
+    public decimal ListPrice { get; set; }
 
     [JsonPropertyName("salePrice")]
-    public double SalePrice { get; set; }
+    public decimal SalePrice { get; set; }
 
     [JsonPropertyName("approved")]
     public bool Approved { get; set; }
